+++ src/tools.js
@@ -3,5 +3,15 @@
 const { createBrowsingTools } = require('../tools/browse');
 const { createMobileTools } = require('../tools/mobile');
 const { createShellTools } = require('../tools/shell');
+const { createSpawnTool, spawnChild } = require('../tools/spawn');
+const { createDeferTool, readQueue: readDeferQueue } = require('../tools/defer');
 
-module.exports = { createBrowsingTools, createMobileTools, createShellTools };
+module.exports = {
+  createBrowsingTools,
+  createMobileTools,
+  createShellTools,
+  createSpawnTool,
+  spawnChild,
+  createDeferTool,
+  readDeferQueue,
+};